Time: 1 ms (99.81%), Space: 43.2 MB (54.9%) - LeetHub

+++ 0389-find-the-difference/0389-find-the-difference.java
@@ -7,7 +7,7 @@ class Solution {
       int xor1=0;
 
       for(int i=0; i<arr1.length; i++){
-        xor1= xor1^ (int)arr1[i];
+        xor1= xor1^ arr1[i];
       }
       
         for(int i=0; i<arr2.length; i++){
